requirements.txt
fastapi
uvicorn
google-generativeai>=0.8.0
feedparser
python-dotenv
requests
pytrends
